connect node

+++ wallethelpers.go
@@ -34,8 +34,8 @@ func mineBlock(t *testing.T, r *coinharness.Harness) {
 	if err != nil {
 		t.Fatal("Failed to GetBestBlock:", err)
 	}
-	count := r.Wallet.Sync(heightAfter)
-	if heightAfter != count {
+	count := r.Wallet.Sync(int64(heightAfter))
+	if int64(heightAfter) != count {
 		t.Fatal("Failed to sync wallet to target:", heightAfter)
 	}
 }
